Put home graphql and styles in pages root
